requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
ariadne==0.13.0
    # via -r requirements.in
arrow==1.1.1
    # via -r requirements.in
asgiref==3.5.0
    # via django
cachetools==5.0.0
    # via django-helusers
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
deprecation==2.1.0
    # via django-helusers
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-crontab
    #   django-db-logger
    #   django-extensions
    #   django-helusers
    #   django-reversion
    #   djangorestframework
    #   drf-oidc-auth
    #   drf-yasg
    #   helsinki-profile-gdpr-api
django-cors-headers==3.11.0
    # via -r requirements.in
django-crontab==0.7.1
    # via -r requirements.in
django-db-logger==0.1.12
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-helusers==0.7.0
    # via
    #   -r requirements.in
    #   helsinki-profile-gdpr-api
django-reversion==4.0.2
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-oidc-auth
    #   drf-yasg
    #   helsinki-profile-gdpr-api
drf-oidc-auth==0.10.0
    # via helsinki-profile-gdpr-api
drf-yasg==1.20.0
    # via -r requirements.in
ecdsa==0.17.0
    # via python-jose
future==0.18.2
    # via pyjwkest
graphql-core==3.2.0
    # via ariadne
gunicorn==20.1.0
    # via -r requirements.in
helsinki-profile-gdpr-api==0.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
packaging==21.3
    # via
    #   deprecation
    #   drf-yasg
psycopg2-binary==2.8.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycryptodomex==3.14.0
    # via pyjwkest
pyjwkest==1.4.2
    # via drf-oidc-auth
pyparsing==3.0.7
    # via packaging
python-dateutil==2.8.2
    # via arrow
python-jose==3.3.0
    # via django-helusers
pytz==2021.3
    # via django
requests==2.27.1
    # via
    #   -r requirements.in
    #   coreapi
    #   django-helusers
    #   pyjwkest
rsa==4.8
    # via python-jose
ruamel.yaml==0.17.20
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
six==1.16.0
    # via
    #   django-db-logger
    #   ecdsa
    #   pyjwkest
    #   python-dateutil
sqlparse==0.4.2
    # via django
starlette==0.14.2
    # via ariadne
typing-extensions==4.0.1
    # via ariadne
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.8
    # via requests
whitenoise==5.2.0
    # via -r requirements.in
xmltodict==0.12.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
